src/main/webapp/app/entities/party-classification/route/party-classification-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { IPartyClassification, PartyClassification } from '../party-classification.model';
import { PartyClassificationService } from '../service/party-classification.service';

import { PartyClassificationRoutingResolveService } from './party-classification-routing-resolve.service';

describe('PartyClassification routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: PartyClassificationRoutingResolveService;
  let service: PartyClassificationService;
  let resultPartyClassification: IPartyClassification | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(PartyClassificationRoutingResolveService);
    service = TestBed.inject(PartyClassificationService);
    resultPartyClassification = undefined;
  });

  describe('resolve', () => {
    it('should return IPartyClassification returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultPartyClassification = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultPartyClassification).toEqual({ id: 123 });
    });

    it('should return new IPartyClassification if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultPartyClassification = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultPartyClassification).toEqual(new PartyClassification());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as PartyClassification })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultPartyClassification = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultPartyClassification).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
